+++ app/build.gradle
@@ -54,6 +54,6 @@ dependencies {
     implementation 'com.squareup.retrofit2:retrofit:2.10.0'
     implementation 'com.google.code.gson:gson:2.11.0'
     implementation project(path: ':okhttpprofiler')
-    implementation project(':requestsModifier')
-//    implementation ('com.localebro:okhttpprofiler:1.0.8')
+    implementation project(':requests-modifier')
+    //    implementation ('com.localebro:okhttpprofiler:1.0.8')
 }
