+++ okhttpprofiler/build.gradle
@@ -45,7 +45,7 @@ dependencies {
     implementation("androidx.compose.ui:ui-tooling")
     
     implementation 'com.google.code.gson:gson:2.11.0'
-    implementation project(':requestsModifier')
+    implementation project(':requests-modifier')
 
     compileOnly 'com.squareup.okhttp3:okhttp:4.9.1'
 }
